backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
firebase-admin==6.6.0
Pillow==11.0.0
ultralytics==8.3.40
opencv-python==4.10.0.84
google-generativeai==0.8.3
requests==2.32.3
twilio==9.3.7